Improved channel index conversion with a handy interface class
This should take care of converting from base 10 readable indices
(as in protocol/script files) to internal handling and vice versa.
There's no sanity check, so use with caution.

+++ src/include/DCThread.h
@@ -39,7 +39,7 @@ private:
 
      inChannel *getInChan(int idx);
      outChannel *getOutChan(int idx);
-     std::vector<std::pair<size_t, bool>> getChanIndices(size_t index);
+     std::vector<std::pair<int, bool> > getChanIndices(int index);
 
      // List of AEC channels (io-channels pairs, kernels and current buffers)
      QVector<AECChannel*> aecChannels;
