+++ src/models/AbHH.cpp
@@ -41,7 +41,7 @@ void abHH::instantiate(std::vector<abHH> &instances, abHHData *inp, DCThread *t,
     tmp.active = true;
     if ( ina->VChannel == ina->IChannel ) {
         // Input/Output on the same model => connect instances 1-to-1 rather than all-to-all
-        for ( std::pair<size_t, bool> VIChan : t->getChanIndices(ina->VChannel) ) {
+        for ( std::pair<int, bool> VIChan : t->getChanIndices(ina->VChannel) ) {
             if ( !VIChan.second )
                 continue;
             tmp.VChannel = VIChan.first;
@@ -49,11 +49,11 @@ void abHH::instantiate(std::vector<abHH> &instances, abHHData *inp, DCThread *t,
             instances.push_back(abHH(inp, t, &tmp));
         }
     } else {
-        for ( std::pair<size_t, bool> VChan : t->getChanIndices(ina->VChannel) ) {
+        for ( std::pair<int, bool> VChan : t->getChanIndices(ina->VChannel) ) {
             if ( !VChan.second )
                 continue;
             tmp.VChannel = VChan.first;
-            for ( std::pair<size_t, bool> IChan : t->getChanIndices(ina->IChannel) ) {
+            for ( std::pair<int, bool> IChan : t->getChanIndices(ina->IChannel) ) {
                 if ( !IChan.second )
                     continue;
                 tmp.IChannel = IChan.first;
